use double precision for urdf importer xform ops

See merge request Isaac/omni_isaac_sim!1929

+++ include/omni/isaac/urdf/Urdf.h
@@ -35,7 +35,7 @@ struct ImportConfig
     UrdfJointTargetType defaultDriveType = UrdfJointTargetType::POSITION;
     float defaultDriveStrength = 1e7f;
     float defaultPositionDriveDamping = 1e5f;
-    float distanceScale = 100.0f;
+    double distanceScale = 100.0;
     UrdfAxis upVector = { 0.0f, 0.0f, 1.0f };
     bool createPhysicsScene = false;
     bool makeDefaultPrim = false;
